+++ src/ioc2rpz_db_sup.erl
@@ -1,4 +1,4 @@
-%Copyright 2017-2019 Vadim Pavlov ioc2rpz[at]gmail[.]com
+%Copyright 2017-2021 Vadim Pavlov ioc2rpz[at]gmail[.]com
 %
 %Licensed under the Apache License, Version 2.0 (the "License");
 %you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ start_db() ->
 %% supervisor.
 init(_Init) ->
 	{ok, []}.
-	
+
 handle_info({'ETS-TRANSFER',Tab,_FromPid,_GiftData}, _State) ->
   ioc2rpz_fun:logMessage("DB_sup got ~p table ownership ~n", [Tab]),
   {noreply, ok};