+++ src/ioc2rpz_rest.erl
@@ -1,4 +1,4 @@
-%Copyright 2017-2019 Vadim Pavlov ioc2rpz[at]gmail[.]com
+%Copyright 2017-2021 Vadim Pavlov ioc2rpz[at]gmail[.]com
 %
 %Licensed under the Apache License, Version 2.0 (the "License");
 %you may not use this file except in compliance with the License.
@@ -25,7 +25,7 @@ init(Req, Opts) ->
     [Op | _] = Opts,
     State = #state{op=Op},
     {cowboy_rest, Req, State}.
-	
+
 allowed_methods(Req, State) ->
     Methods = [<<"GET">>, <<"POST">>],
     {Methods, Req, State}.
@@ -33,7 +33,7 @@ allowed_methods(Req, State) ->
 content_types_provided(Req, State) ->
     {[
       {<<"application/json">>, to_json},
-      {<<"text/plain">>, to_txt}	  
+      {<<"text/plain">>, to_txt}
      ], Req, State}.
 
 is_authorized(Req, State) ->
@@ -41,7 +41,7 @@ is_authorized(Req, State) ->
 	[[MKeysT,ACL,Srv]] = ets:match(cfg_table,{srv,'_','_','$4','$5','_','$7'}),
 	MKeys=lists:flatten([ MKeysT,[ ets:match(cfg_table,{[key_group,X,'_'],'$3'}) || X <- Srv#srv.key_groups ] ]),
 
-	MGMTIP=ioc2rpz_fun:ip_in_list(ioc2rpz:ip_to_str(IP),ACL),	
+	MGMTIP=ioc2rpz_fun:ip_in_list(ioc2rpz:ip_to_str(IP),ACL),
 
 	case {cowboy_req:parse_header(<<"authorization">>, Req),MGMTIP} of
 		{{basic, User, Password}, true} ->
@@ -59,7 +59,7 @@ is_authorized(Req, State) ->
 			end;
 		{{bearer, Token}, true} ->
             ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(131),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
-			{false, Req, State#state{user=Token}};		
+			{false, Req, State#state{user=Token}};
 
 		{_, false} ->
             ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(135),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
@@ -72,7 +72,7 @@ is_authorized(Req, State) ->
 			Req0=cowboy_req:set_resp_body(Body,Req),
 			{{false, <<"Basic">>}, Req0, State}
 	end.
-	
+
 
 to_json(Req, State) ->
 	srv_mgmt(Req, State, json).
@@ -105,6 +105,45 @@ srv_mgmt(Req, State, Format) when State#state.op == update_tkeys -> %Reload TSIG
 	end,
 	{Body, Req0, State};
 
+srv_mgmt(Req, State, Format) when State#state.op == cache_sources_clear_all -> % clear all sources from the hot cache
+	#{peer := {IP, Port}} = Req,
+  ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(230),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
+  %SW=ets:match(cfg_table, {[source,'_'],'$2'}),
+  %[ ioc2rpz_fun:logMessage("deleting ~p source~n",[X#source.name]) || [X] <- SW ],
+  %[ ets:delete(rpz_hotcache_table, {X#source.name,Y}) || [X] <- SW, Y <-[axfr,ixfr] ],
+  [ ioc2rpz_fun:logMessage("deleting ~p ~p source from the hotcache~n",[X,Y]) || [X,Y] <- ets:match(rpz_hotcache_table,{{'$1','$2'},'_','_'})],
+  [ ets:delete(rpz_hotcache_table, {X,Y}) || [X,Y] <- ets:match(rpz_hotcache_table,{{'$1','$2'},'_','_'})],
+	Body = case Format of
+		json -> "{\"status\":\"ok\",\"msg\":\"All sources were removed from the hotcache\"}\n";
+		txt -> "status: ok\nmsg: All sources were removed from the hotcache\n"
+	end,
+	{Body, Req, State};
+
+srv_mgmt(Req, State, Format) when State#state.op == cache_sources_clear_one-> % clear a source from the hot cache
+	#{peer := {IP, Port}} = Req,
+  ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(230),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
+  Source = binary_to_list(cowboy_req:binding(source, Req)),
+  %[ ioc2rpz_fun:logMessage("~p ~p source in cache~n",[X,Y]) || [X,Y] <- ets:match(rpz_hotcache_table,{{'$1','$2'},'_','_'})],
+  ioc2rpz_fun:logMessage("deleting ~p source from the hot cache~n",[Source]),
+  ets:delete(rpz_hotcache_table, {Source,axfr}), ets:delete(rpz_hotcache_table, {Source,ixfr}),
+  %[ ioc2rpz_fun:logMessage("~p ~p source in cache~n",[X,Y]) || [X,Y] <- ets:match(rpz_hotcache_table,{{'$1','$2'},'_','_'})],
+	Body = case Format of
+		json -> io_lib:format("{\"status\":\"ok\",\"msg\":\"~s source was removed from the hot cache\"}\n",[Source]);
+		txt -> io_lib:format("status: ok\nmsg: ~s source was removed from the hot cache\n",[Source])
+	end,
+	{Body, Req, State};
+
+srv_mgmt(Req, State, Format) when State#state.op == cache_sources_load_all -> % load all sources to the hotcache
+	#{peer := {IP, Port}} = Req,
+  ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(230),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
+  SW=[X#source.name || [X] <- ets:match(cfg_table, {[source,'_'],'$2'})],
+  %ioc2rpz:mrpz_from_ioc(SW,#rpz{serial=ioc2rpz_fun:curr_serial()},axfr,[]),
+  spawn_opt(ioc2rpz,mrpz_from_ioc,[SW,#rpz{serial=ioc2rpz_fun:curr_serial()},axfr,[]],[{fullsweep_after,0}]),
+	Body = case Format of
+		json -> "{\"status\":\"ok\",\"msg\":\"All sources will loaded to the hot cache\"}\n";
+		txt -> "status: ok\nmsg: All sources will loaded to the hot cache\n"
+	end,
+	{Body, Req, State};
 
 srv_mgmt(Req, State, Format) when State#state.op == update_all_rpz -> % Force update all RPZ zones
 	#{peer := {IP, Port}} = Req,
@@ -142,7 +181,7 @@ srv_mgmt(Req, State, Format) when State#state.op == terminate -> %Shutdown serve
 	end,
 	ioc2rpz_sup:stop_ioc2rpz_sup(),
 	{Body, Req, State};
-		
+
 srv_mgmt(Req, State, Format) when State#state.op == stats_serv -> % Statistics -- TODO
 	#{peer := {IP, Port}} = Req,
     ioc2rpz_fun:logMessageCEF(ioc2rpz_fun:msg_CEF(230),[ioc2rpz:ip_to_str(IP), Port, cowboy_req:path(Req), ""]),
@@ -202,7 +241,7 @@ srv_mgmt(Req, State, Format) when State#state.op == get_rpz -> % Get RPZ
 % ioc2rpz_db:lookup_db_record(<<"baddomain1.com">>,no).
 % ioc2rpz_db:lookup_db_record(<<"99.98.61.5">>,no).
 % DB
-%11> ets:select(rpz_ixfr_table,[{{{ioc,'$0',<<"99.98.61.5">>},'$2','$3'},[],[{{'$0','$2','$3'}}]}]).   
+%11> ets:select(rpz_ixfr_table,[{{{ioc,'$0',<<"99.98.61.5">>},'$2','$3'},[],[{{'$0','$2','$3'}}]}]).
 %[{<<5,108,111,99,97,108,7,105,111,99,50,114,112,122,0>>,
 %  1572419220,0},
 % {<<8,108,111,99,97,108,45,105,112,7,105,111,99,50,114,112,
@@ -245,8 +284,8 @@ srv_mgmt(Req, State, Format) when State#state.op == catch_all -> % Catch all uns
 %		txt ->  cowboy_req:reply(501,#{<<"content-type">> => <<"text/html">>}, ["status: error\nmsg: Unsupported request\n"],Req)
 %    end,
 %	{false, Req0, State}.
-	
-rest_terminate(Req, State) ->
+
+rest_terminate(_Req, _State) ->
 	ok.
 
 format_ioc({ok,Results},Req,Format) ->
@@ -264,11 +303,11 @@ format_ioc([],{IOC,TKEY,_Zones},json, Result) ->
 format_ioc([{El,Feeds}|Results],Req,json,"") ->
 	Ind=io_lib:format("{\"ioc\": \"~s\", \"feeds\": ~s}",[El, parse_feeds(Feeds,Req,"",json)]),
 	format_ioc(Results,Req,json, Ind);
- 
-format_ioc([{El,Feeds}|Results],Req,Format,Result) ->
+
+format_ioc([{El,Feeds}|Results],Req,_Format,Result) ->
 	Ind=io_lib:format("{\"ioc\": \"~s\", \"feeds\": ~s}",[El, parse_feeds(Feeds,Req,"",json)]),
 	format_ioc(Results,Req,json, Result ++","++ Ind).
-	
+
 
 parse_feeds([],_Req,Result,json) ->
 	"["++Result++"]";
@@ -277,7 +316,7 @@ parse_feeds([{Feed, Serial, Exp}|REST],{_IOC,_TKEY,Zones}=Req,"",json) ->
 	Memb=maps:is_key(Feed,Zones),
 	Feed_Str=if (Memb) -> {FN,TY,WC}=maps:get(Feed,Zones), io_lib:format("{\"feed\":~p, \"wildcard\":~s, \"type\":~p, \"rpz_serial\": ~p, \"ioc_expiration\": ~p}",[FN, WC, binary_to_list(TY), Serial, Exp]); true -> "" end,
 	parse_feeds(REST,Req,Feed_Str,json);
-	
+
 parse_feeds([{Feed, Serial, Exp}|REST],{_IOC,_TKEY,Zones}=Req,Result,json) ->
 	Memb=maps:is_key(Feed,Zones),
 	Feed_Str=if (Memb) -> {FN,TY,WC}=maps:get(Feed,Zones), ","++io_lib:format("{\"feed\":~p, \"wildcard\":~s, \"type\":~p, \"rpz_serial\": ~p, \"ioc_expiration\": ~p}",[FN, WC, binary_to_list(TY), Serial, Exp]); true -> "" end,
@@ -288,14 +327,14 @@ parse_feeds([{Feed, Serial, Exp}|REST],{_IOC,_TKEY,Zones}=Req,Result,json) ->
 %%%
 get_tkey_zones(TKey) ->
 	{ok, TKeyBin} = ioc2rpz:domstr_to_bin(TKey,0),
-	Groups = [ X || [X,Y] <- ets:match(cfg_table,{[key_group,'$1',TKeyBin],'$3'}) ],
+	Groups = [ X || [X,_Y] <- ets:match(cfg_table,{[key_group,'$1',TKeyBin],'$3'}) ],
 	get_tkey_zones(TKeyBin, Groups, [ X || [X] <- ets:match(cfg_table,{[rpz,'_'],'_','$4'}) ], []). %{X#rpz.zone, X#rpz.zone_str, X#rpz.wildcards, X#rpz.akeys, X#rpz.ioc_type, X#rpz.key_groups}
 
-get_tkey_zones(TKeyBin, _Groups,[], Zones) ->
+get_tkey_zones(_TKeyBin, _Groups,[], Zones) ->
 	Recur = [ X || {_,{_,_,X}} <- Zones, X == <<"true">> ] /= [],
 %	ZNames = [ X || {X,{_,_,_}} <- Zones ],
 	{Recur, maps:from_list(lists:flatten(Zones))};
-	
+
 get_tkey_zones(TKeyBin, Groups, [RPZ|Rest], Zones) ->
 	KZ = lists:member(TKeyBin, RPZ#rpz.akeys),
 	GZ = [X || X <- Groups, lists:member(X,RPZ#rpz.key_groups)],
@@ -321,14 +360,14 @@ gen_srv_stats(Format) ->
   MemAXFR = binary_to_list(ioc2rpz_fun:conv_to_Mb(ioc2rpz_db:db_table_info(rpz_axfr_table,memory) * WS)),
   MemIXFR = binary_to_list(ioc2rpz_fun:conv_to_Mb(ioc2rpz_db:db_table_info(rpz_ixfr_table,memory) * WS)),
 	gen_srv_stats(Format,[Node,Srv_rules,MemHC,MemAXFR,MemIXFR]).
- 
+
 gen_srv_stats(txt, [Node,Srv_rules,MemHC,MemAXFR,MemIXFR]) ->
   io_lib:format("node_name ~p\n srv_total_rules ~b\n hot_cache_mem ~s\n axfr_table_mem ~s\n ixfr_table_mem ~s\n",[Node,Srv_rules,MemHC,MemAXFR,MemIXFR]);
 gen_srv_stats(json, [Node,Srv_rules,MemHC,MemAXFR,MemIXFR]) ->
   io_lib:format("{\"node_name\":\"~p\",\"srv_total_rules\":~b,\"hot_cache_mem\":\"~s\",\"axfr_table_mem\":\"~s\",\"ixfr_table_mem\":\"~s\"}",[Node,Srv_rules,MemHC,MemAXFR,MemIXFR]).
 
 list_tuples_to_json(Array) ->
-    io_lib:format("[~s]",[list_tuples_to_json([],Array)]).    
+    io_lib:format("[~s]",[list_tuples_to_json([],Array)]).
 
 list_tuples_to_json([],[E|Rest]) ->
     list_tuples_to_json(tuple_to_json(E),Rest);
@@ -341,7 +380,7 @@ list_tuples_to_json(Resp,[]) ->
 
 tuple_to_json({Name,Value}) when is_integer(Value)->
     io_lib:format("{\"~s\":~b}",[Name,Value]);
-    
+
 tuple_to_json({Name,Value}) ->
     io_lib:format("{\"~s\":\"~s\"}",[Name,Value]);
 
@@ -351,19 +390,19 @@ tuple_to_json(REST) ->
 
 mtuple_to_json([],[{Name,Value}|REST]) when is_integer(Value)->
     mtuple_to_json(io_lib:format("\"~s\":~b",[Name,Value]),REST);
-    
+
 mtuple_to_json([],[{Name,Value}|REST]) ->
     mtuple_to_json(io_lib:format("\"~s\":\"~s\"",[Name,Value]),REST);
 
 mtuple_to_json(Val,[{Name,Value}|REST]) when is_integer(Value)->
     mtuple_to_json(Val++io_lib:format(",\"~s\":~b",[Name,Value]),REST);
-    
+
 mtuple_to_json(Val,[{Name,Value}|REST]) ->
     mtuple_to_json(Val++io_lib:format(",\"~s\":\"~s\"",[Name,Value]),REST);
 
 mtuple_to_json(Val,[]) ->
     Val.
-		
+
 ioc2jsonarr(IOCs) ->
     %ioc2rpz_fun:logMessage("~p\n\n",[IOCs]),
     ioc2jsonarr([],IOCs).